# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Key Stone","description":"Key Stone is a platform for managing your keys","version":"v0.0.1"},"host":"localhost:80","basePath":"/key-stone","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/auth":{"post":{"tags":["token"],"summary":"issue token","operationId":"token#issue","parameters":[{"name":"IssueRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/TokenIssueRequestBody","required":["user"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/TokenDetail","required":["access_token","token_type","expires_in","refresh_token"]}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TokenIssueUnauthorizedResponseBody"}}},"schemes":["http"]}},"/auth/refresh":{"post":{"tags":["token"],"summary":"refresh token","operationId":"token#refresh","parameters":[{"name":"RefreshRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/TokenRefreshRequestBody","required":["token"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/TokenDetail","required":["access_token","token_type","expires_in","refresh_token"]}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TokenRefreshUnauthorizedResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/TokenRefreshInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/users":{"post":{"tags":["user"],"summary":"create user","operationId":"user#create","parameters":[{"name":"CreateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserCreateRequestBody","required":["user"]}}],"responses":{"204":{"description":"No Content response."}},"schemes":["http"]}},"/users/me":{"delete":{"tags":["user"],"summary":"delete user","operationId":"user#delete","parameters":[{"name":"Authorization","in":"header","description":"The authorization header","required":true,"type":"string"}],"responses":{"204":{"description":"No Content response."},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserDeleteUnauthorizedResponseBody"}}},"schemes":["http"]}}},"definitions":{"IssueInput":{"title":"IssueInput","type":"object","properties":{"password":{"type":"string","description":"The password of the user","example":"Tempora numquam nisi eaque."},"username":{"type":"string","description":"The username of the user","example":"Et excepturi eligendi nihil consequatur incidunt."}},"example":{"password":"Suscipit non ad adipisci accusamus.","username":"Explicabo ipsum distinctio consequatur consectetur blanditiis repudiandae."},"required":["username","password"]},"RefreshInput":{"title":"RefreshInput","type":"object","properties":{"access_token":{"type":"string","description":"The access token of the user","example":"Laborum illum tempore magnam."},"refresh_token":{"type":"string","description":"The refresh token of the user","example":"Qui est."}},"example":{"access_token":"Consequatur illum ducimus reprehenderit assumenda.","refresh_token":"Rerum fugit sed nemo."},"required":["access_token","refresh_token"]},"TokenDetail":{"title":"TokenDetail","type":"object","properties":{"access_token":{"type":"string","description":"The access token of the user","example":"Reprehenderit occaecati qui in."},"expires_in":{"type":"integer","description":"The expires in of the user","example":7100820663025279588,"format":"int64"},"refresh_token":{"type":"string","description":"The refresh token of the user","example":"Officia vel dolorem fuga vel omnis neque."},"token_type":{"type":"string","description":"The token type of the user","example":"Repellat voluptatem consectetur."}},"example":{"access_token":"Animi ut nihil id alias culpa optio.","expires_in":2627895758942114620,"refresh_token":"Nihil impedit similique autem.","token_type":"Voluptatem accusantium minus repellendus non."},"required":["access_token","token_type","expires_in","refresh_token"]},"TokenIssueRequestBody":{"title":"TokenIssueRequestBody","type":"object","properties":{"user":{"$ref":"#/definitions/IssueInput"}},"example":{"user":{"password":"Dolores fuga asperiores voluptatem aspernatur eum.","username":"Explicabo placeat."}},"required":["user"]},"TokenIssueUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Unauthorized (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TokenRefreshInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Internal Server Error (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TokenRefreshRequestBody":{"title":"TokenRefreshRequestBody","type":"object","properties":{"token":{"$ref":"#/definitions/RefreshInput"}},"example":{"token":{"access_token":"Et quia illum.","refresh_token":"Distinctio quam sunt est sed alias."}},"required":["token"]},"TokenRefreshUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Unauthorized (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"UserCreateRequestBody":{"title":"UserCreateRequestBody","type":"object","properties":{"user":{"$ref":"#/definitions/UserInput"}},"example":{"user":{"password":"Occaecati vel adipisci.","username":"Quia architecto veritatis deleniti ea perferendis."}},"required":["user"]},"UserDeleteUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Unauthorized (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"UserInput":{"title":"UserInput","type":"object","properties":{"password":{"type":"string","description":"The password of the user","example":"Cumque suscipit quia aperiam tempore."},"username":{"type":"string","description":"The name of the user","example":"Minus cumque quia."}},"example":{"password":"Non non qui repellendus.","username":"Dolore vero et voluptatem eius."},"required":["username","password"]}}}